Known Working Master Code

+++ Master-Node/master/master.ino
@@ -2,6 +2,7 @@
  This will act as the boombox controller. It will balance transmitting out to the nodes and receiving their data, instead of only doing 1 function
 */
 
+
 #define ENABLE false
 
 #include <Wire.h>
@@ -53,8 +54,7 @@ void loop() {
   while ( network.available() ) {     // Is there any incoming data?
     //Serial.println("RX");
     RF24NetworkHeader header;
-    //int incomingData;
-    uint8_t incomingData;
+    int incomingData;
     network.read(header, &incomingData, sizeof(incomingData)); // Read the incoming data
     
     if(header.from_node == 01 ){ //TAKE ACTION ON RECIEVE PAYLOAD
